move util fns into expects mod
